+++ lib/recording_studio_youtube.rb
@@ -17,7 +17,6 @@ require "recording_studio_youtube/session"
 require "recording_studio_youtube/ai_tools"
 require "recording_studio_youtube/diagnostics"
 require "recording_studio_youtube/engine"
-require "recording_studio_youtube/capabilities/example"
 
 module RecordingStudio
   module YouTube